msm: vidc: Decrease lower threshold by extra buffers needed for DCVS
The lower threshold is a rough estimate for the number of buffers that
the client is likely to be holding.  Previously we considered that to be
the total number of buffers - minimum required by the firmware.

The total number of buffers includes extra buffers required for DCVS
which the client will most likely not hold.  So subtract that out from
the lower threshold as well.

With this change, the nominal residency is increased drastically.

Change-Id: I328bc0a7e8c52579225c09bd572cfbcbf99c4070
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>